PR feedback (consolidating ref files, ...)
