+++ src/System.ValueTuple/ref/System.ValueTuple.csproj
@@ -15,8 +15,6 @@
   </PropertyGroup>
   <ItemGroup Condition="'$(TargetGroup)' != ''">
     <Compile Include="System.ValueTuple.cs" />
-    <Compile Include="TupleElementNamesAttribute.cs" />
-    <Compile Include="TupleExtensions.cs" />
   </ItemGroup>
   <ItemGroup Condition="'$(TargetGroup)' == 'portable-net40+sl4+win8+wp8'">
     <!-- Those files only serve to unblock the compilation given that this profile didn't have those attributes -->
