--- src/System.ValueTuple/ref/TupleElementNamesAttribute.cs
@@ -1,19 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-// ------------------------------------------------------------------------------
-// Changes to this file must follow the http://aka.ms/api-review process.
-// ------------------------------------------------------------------------------
-
-using System.Collections.Generic;
-
-namespace System.Runtime.CompilerServices
-{
-    [CLSCompliant(false)]
-    [AttributeUsage(AttributeTargets.Field | AttributeTargets.Parameter | AttributeTargets.Property | AttributeTargets.ReturnValue | AttributeTargets.Class | AttributeTargets.Struct | AttributeTargets.Event )]
-    public sealed class TupleElementNamesAttribute : Attribute
-    {
-        public TupleElementNamesAttribute(string[] transformNames) { throw null; }
-        public IList<string> TransformNames { get { throw null; } }
-    }
-}
