+++ src/System.ValueTuple/ref/project.json
@@ -15,7 +15,7 @@
     },
     "netcoreapp1.1": {
       "dependencies": {
-        "System.Runtime": "4.4.0-beta-24905-02"
+        "System.Runtime": "4.4.0-beta-24906-02"
       }
     }
   }
